Migrate Client ParserOutputUpdateHookHandler hook to extension JSON
Like many other Client hooks, we migrate this one to the new hook
container system, but unlike in the other cases, we keep the distinction
between the on…() and do…() methods. The on…() method skips the hook
during unit tests, so the do…() method can be called directly by the
hook’s own unit tests; it looks like skipping the hook during unit tests
does make some sense, so for now I think we should leave it like that.

Bug: T256764
Change-Id: Ifd3a339d973e95e3de171b81ca47954775248d30