Don’t load $wgWBClientSettings in WikibaseClient.php
Previously, WikibaseClient.php would populate the $wgWBClientSettings
global with the default settings, then LocalSettings.php would override
parts of it, and then later WikibaseClient::newInstance() would call
WikibaseSettings::getClientSettings() to turn that array into a
SettingsArray (after normalizing the entityNamespaces setting).

Now, we don’t want to load $wgWBClientSettings in the PHP entry point
anymore (because we want to turn that into an extension JSON file, and
there’s no way to run PHP code as soon as that file is queued using
wfLoadExtension() – the whole file is only processed later, when the
rest of LocalSettings.php would already have run). So instead, we only
load the default settings in WikibaseSettings::getClientSettings(). At
this point, LocalSettings.php has already partially populated the
$wgWBClientSettings global (silently creating it with the first
assignment, because that’s how PHP array assignments work), so now we
need to carefully merge the default settings into it so that no
difference is visible: the custom settings should override the default
settings, but if the default is an array, then we assume that the custom
settings were assigned using $wgWBClientSettings['name'][] = 'append' or
$wgWBClientSettings['name']['subkey'] = 'value' syntax, and the defaults
should partially be kept.

Bug: T256252
Change-Id: I193008a732798b9b73ad11faf61e448baef77d6a