lib: Use void for hook handlers that should never be aborted
For these there is no intent that WikibaseLib might want to
prevent other extensions from using this hook.

Not all of these are technically required by core to be void,
but it's harmless to proactively mark them as void here. This
helps with static analysis to automatically flag if you tried to
return something from here.

Change-Id: Ie8b054de0867826eb59821c143361f7243757ed7