Simplify GlobalStateFactoryMethodsResourceTest
Since we already record ObjectFactory-compatible specifications in
extension.json, we can simplify GlobalStateFactoryMethodsResourceTest:
instead of hard-coding one test per hook handler, API module, or
similar, iterate over all the specifications recorded in the extension
JSON file and test each of them.

The relationships between the test functions and their data providers
may look a bit strange: the data providers have both the name and the
corresponding specification, but only return the name, and the test
functions then look up the specifications again. The motivation for this
is to keep the test names reported by PHPUnit simple – I don’t think
it’s useful to see the whole specification array when one of these tests
fails, the name should be enough.

Change-Id: Id56fc6a0a41188d40bc9bb09a6a9ff1c8bb1c8be