Don’t use WB_VERSION in Client example config
The WikibaseClient sample config (defined in WikibaseClient.example.php,
but more commonly included via client/ExampleSettings.php) defines a
sample repo for this client, but only if this wiki is not also the repo,
because in that case the default config is already sufficient. I expect
that eventually we’ll get rid of this entirely, and require non-repo
clients to explicitly configure the repo (the example repo is of limited
usefulness anyways), but for now, we only replace the use of one
constant with another one: instead of WB_VERSION, which we want to
remove because it’s too general and shouldn’t be used, we introduce a
very specific constant that’s only used for this one communication
between repo/Wikibase.php and client/config/WikibaseClient.example.php,
and then only temporarily (hopefully).

Bug: T256238
Change-Id: Ie42d0218c286a8783b8cde9349b69c4bfc608f04