Rewrite LuaWikibaseIntegrationTest as API integration test
We want to avoid the direct coupling to the WikibaseRepo class from this
Client test, and chose to do that by rewriting it as an API integration
test [1]. Most of the test characteristics can be retained (using API
calls to create items, parse wikitext, and get entity usage), but since
we have less control over the target wiki, we can’t port the getSitelink
test (we don’t know any valid sitelink code – not even the repo wiki’s
own code is guaranteed to be registered), and some tests have to be more
lenient as to which result language they expect.

Since the API integration tests aren’t voting in CI yet, this change
doesn’t remove the PHP version of the test, but that will follow soon.

[1]: https://www.mediawiki.org/wiki/MediaWiki_API_integration_tests

Bug: T255885
Change-Id: I62c520f8269f304b492390760d139c2ceea11752