Decouple jquery.ui.suggester RL module from client and repo
Copying the module from lib to client and use it directly in linkitem.
Moving the file from lib/ to view/ and register it as a module there
instead.

Also running automatic fixes on eslint and stylelint and some manual
changes to fix errors that couldn't be fixed automatically.

And last but not least, changing colors to align with the WikimediaUI
color palette which was already decided and partially implemented: I5320e3b9984
This change is almost invisible to naked eye:
https://design.wikimedia.org/style-guide/visual-style_colors.html

Bug: T257476
Change-Id: Ifcc0cd5c0f16b9d72bf71d264b033690a7240b09