Decouple jquery.event.special.eachchange and util.inherit
Copying the module from lib to client and use it directly in linkitem.
Moving the file from lib/ to view/ and register it as a module there
instead.

Also running automatic fixes on eslint and some manual
changes to fix errors that couldn't be fixed automatically and adding
ignore rules for the ones that are not easy to fix.

This is last pieces to finish the decoupling linkitem.

Bug: T257452
Bug: T256054
Change-Id: Icad3c338cf217f2f940e332bed19b1b7807d5fdd