Add entityterms Repo API
This does pretty much the same thing that the pageterms Client API
currently does on Repo wikis: return terms of the entity in the given
page. However, we want to remove that item from pageterms, hence the
dedicated API module.

I considered refactoring the existing PageTerms class so that it could
be reused for entityterms, but decided against it. The refactored class
would have had to go into Lib, but we want to avoid adding code there
(and Lib doesn’t currently hold any APIs), and yet the refactored class
makes little sense as a separate library either; and we also don’t
expect any further instantiations of the abstract API beyond pageterms
and entityterms, so a copy+paste of the code seems acceptable to me.
(But while we’re at it, let’s also modernize the PageTerms code a bit.)

The message apihelp-query+pageterms-example-item is repurposed as an
example message for the Repo API instead, and consequently renamed to
apihelp-query+entityterms-example-item.

Bug: T257658
Change-Id: Iab514e4e8a10022fe7dce52ab3a44a5410334ddd