Fix limit not applied in case of language fallback
Added a new test that covers the fallback limit. The tests are now
somewhat brittle as they depend on a particular ordering of the test
data.  Unfortunately, that seems unavoidable as it has to be ensured
that sufficient fallback data comes before the data with the initially
specified language in order to saturate the request for fallback data.
Because only then the limit in production code is actually applied and
thus covered by the test.

The mock was changed as it didn't mock the interface correctly, in
particular when both description and label match and the filtering down
to only 1 result per entity happens _after_ the limit on the total
results was applied. This made debugging harder even though it doesn't
affect the currently existing tests.

Bug: T256885
Change-Id: Ia2563b52f2faeac03664b2a6c0821ee5f0996e39