Rename extension-repo-wip.json to extension-repo.json
The file is not ready for standalone use yet, but we would like to start
referencing it in production config soon, e. g. in the extension-list
(which is only used to discover the i18n globals, and those we already
migrated), and that works better if the file already has its final name.

Change-Id: Ia3df07de4dd78d9bde52fbe080d2797efceeebde