Migrate wikibase.api module to repo and use it directly in client
It's used only in one module in client, instead of making another module
we can just move it to repo and call it directly in linkitem module in
client.

Note that unlike other decoupling, we haven't copied the js code, they
still exist in one shared place (a git submodule) so they are still
coupled to some degree but not on RL module level.

Bug: T256054
Change-Id: Iad6df9ff518edf8a11f8dbed71c1d609b2dd743a