TR: storybook: update & fix warning
Updates storybook and friends.

Also fixes a pre-existing warning about the failure to register an addon
which is not installed (knobs) - it is not used, so I disabled the
registration for now.

There are a few things that remain to be done:
* storybook-addon-vue-info, which is kind of deprecated, should be
  exchanged for the docs-addon
* addon registration could should also be consolidated to one way of
  doing it.
* depending on some feedback by the orginal authors the mention of knobs
  should be removed entirely or the planned knobs implemented

Deferring to do more updates instead (and gather feedback from experts).

Change-Id: I34f228c610c437e10a3dd4caa262524e34d41d1b