Don’t load random strings as classes in SummaryFormatter
On my system, SetAliasesTest::testSetAliases_cannotCreateProperty
failed:

> Wikibase\Repo\SummaryFormatter::formatArg: failed to render value:
> Methods with the same name as their class will not be constructors in a
> future version of PHP; Foo has a deprecated constructor [Called from
> Wikibase\Repo\SummaryFormatter::formatArg in
> /var/www/html/wiki1/extensions/Wikibase/repo/includes/SummaryFormatter.php
> at line 198]

This is because the test eventually caused SummaryFormatter to call
method_exists( 'Foo', '__toString'), which made PHP try to load a Foo
class, and it actually found one:

    $ cat vendor/php-parallel-lint/php-parallel-lint/tests/examples/example-05/Foo.php
    <?php

    class Foo
    {

        /**
         * @var string
         */
        private $bar;

        public function Foo($bar)
        {
            $this->bar = $bar;
        }
    }

This class has a malformed PHP constructor, causing that warning. I
assume that SummaryFormatter only meant to test if the *object* $arg has
a __toString method, not if a class with the name of the *string* $arg
has one, so move this check inside the block guarded by is_object().

Change-Id: I39f80ee2b6e37dc1f897dccb722489fa672b3d90