Merge ChangesSubscriptionSchemaUpdater into DatabaseSchemaUpdater
Currently, we have two hook handlers for LoadExtensionSchemaUpdates hook
and this is not a problem on its own but
ChangesSubscriptionSchemaUpdater hook handler actually depends on
DatabaseSchemaUpdater one to be ran first, otherwise it explodes
majestically and won't get installed in fresh setups because it can't
find the table to alter. This is a implicit logic hidden in the
Wikibase.php which is suboptimal.

The simple and proper solution is to merge the second one into the main
database schema updater. `wb_changes_subscription` is not in any way
more special than other Wikibase tables and they can and should have
their update logic in one class to reduce unnecessary complexity.

Bug: T257563
Change-Id: Ifab4b19dba85af546254f887aaf444150946cb3b