TR: vuex: update (to version found in MW)
Update the vuex dependency, but no further than what MediaWiki can
provide[0] once the vuex dependency is externalized during T250267.

We do not want our dev version to overtake what is available in
production. Sure, it would appease npm audit and give us new features in
dev but we would not be able to benefit from them in production and - in
the worst case - cause a problem which we would only detect there.

If we one day feel the need to update what is externalized, our energy
should be _first_ directed to core to update its version of the
externalized package, _then_ bump it to that version here.

[0]: https://github.com/wikimedia/mediawiki/blob/00192a6/resources/lib/vuex/vuex.js#L2

Change-Id: I3dc08852edb5a0e8b0a5e98c83153eff51fabb1d