Ensure PSR-16 valid cache keys
PSR-16 explicitly states a number of invalid characters for cache keys:

> The following characters are reserved for future extensions and
> MUST NOT be supported by implementing libraries: {}()/\@:

This caused an exception in production when used with a Lexeme key
similar to  "L:L123". Even though that input is probably the result of a
bug somewhere else, this method here should not be prone to creating
invalid cache keys either way.

Change-Id: I2573601e0462550da052cb24597f5b9137ff5db3