Make onResourceLoaderRegisterModules less general
The method only registers a single ResourceLoader module now, and
completely skips itself if that one module is already registered, so I
don’t want the rest of the method to look like more modules could be
added there. Rearrange the code so that it’s more clearly only for
wikibase.Site and nothing else.

Change-Id: Iab8555c77e41910a15ab6887c9aebdfbd8f0aba4