Inject LanguageConverterFactory into SpecialEntityUsage
Previously, the special page got the ContentLanguage from the service
container, but it only used the content language for the deprecated
convert() method, for which we should use the newer LanguageConverter
class instead. Inject a LanguageConverterFactory so that we can create
the right LanguageConverter on demand (the content language is the
default for getLanguageConverter() without arguments).

And since I have to touch SpecialEntityUsageTest::newSpecialPage()
anyways, slightly improve the mocking code there as well (setMethods()
is soft deprecated, and the method mocking was more complicated than it
needed to be IMHO).

Change-Id: I51198b1cbe6729bdc2f89904c7ee5ae57c0b423a