Rename LanguageFallbackChain because it is about Terms
The LanguageFallbackChain class was source for confusion whether it was
intended (and safe) to be used for Terms or whether might also be used
for user interface message considerations.

This patch now renames the class to make its intend obvious. Follow-up
patches will include validation to ensure that only valid term languages
are included.

The alias will be removed in Ib20eef3d75400a65b98cbfe28dd61b146d1adc31.

Bug: T250930
Change-Id: Ic31b720669601b78f034a80c3337eb987f4c743e