Don’t configure example repo in client example config
There’s no way to make this work without the PHP entry point, so we
announced to third-party users that they may need to update their
Wikibase configuration, and disable the block completely. (In
production, the example config isn’t used anyways.)

Since the example config is now a no-op (the two settings it still sets
to true are also true by default), stop loading it in the CI config.
Instead, the CI config now needs to configure some Repo for Client-only
tests; fortunately, in this case, we actually have a global variable
(set in build/jenkins/mw-apply-wb-settings.sh) that we can use to detect
this case.

Bug: T257449
Change-Id: Iebdd530d3ba01a47feb453b985f4221fc1fe34a3