changes: Define MW_PHPUNIT_TEST if necessary
EntityChange::getSerializedInfo() is supposed to never serialize
objects, but because the check for that is expensive, that check is only
enabled in the unit tests. Define the necessary constant so that the
test still works in the standalone library. (In MediaWiki, the constant
is simply true; here, set it to a string value that hopefully explains
what’s happening if anybody happens to look at the constant value.)

Change-Id: Ia8905266c6af0b87cbf7cd634833faead6a77038