Don’t load Repo settings in PHP entry point
Previously, Wikibase.php would populate the $wgWBRepoSettings
global with the default settings, then LocalSettings.php would override
parts of it, and then later WikibaseRepo::newInstance() would call
WikibaseSettings::getRepoSettings() to turn that array into a
SettingsArray (after normalizing the entityNamespaces setting).

Now, we don’t want to load $wgWBRepoSettings in the PHP entry point
anymore (because we want to turn that into an extension JSON file, and
there’s no way to run PHP code as soon as that file is queued using
wfLoadExtension() – the whole file is only processed later, when the
rest of LocalSettings.php would already have run). So instead, we only
load the default settings in WikibaseSettings::getRepoSettings(). At
this point, LocalSettings.php has already partially populated the
$wgWBRepoSettings global (silently creating it with the first
assignment, because that’s how PHP array assignments work), so now we
need to carefully merge the default settings into it so that no
difference is visible.

For most cases, it's easy but in some configs it can get tricky, e.g.
these two lines in LocalSettings.php are both valid
$wgWBRepoSettings['urlSchemes'][] = 'gopher';
and
$wgWBRepoSettings['urlSchemes'] = [ 'gopher' ];

For values that are not arrays, it's simple, just override but for
non-empty arrays, we need to define three merge strategies and apply
them per config:
 - Use "wfArrayPlus2d": like "string-limits"
 - "false" means remove: If the array sets a key to false, we should
   just remove that key.
 - Override (i.e. do not merge): Like entityDataFormats

OF course, this is hacky and we should migrate to mediawiki's config
management. Filed T258658 for detailed and long-term solution.

Bug: T257447
Depends-On: Iffe7e5e05581c85ed593d73293f001b1c89a2316
Change-Id: Iebcb8495c1861277edfdbe6c10a09e6f92e3c94a