FP: use the local ns id for properties in entitysources
With the way our existing services are currently wired up etc
this ns ID needs to match the local location that properties would
be if they were local, so that links etc get created with the
correct prefix.

If this is set to the namespace ID of the remote wiki,
the links end up wrong, for example Item:P31.

A concrete example would be:
 My local dev install, NS120 = items, NS 122 = properties
 Wikidata, NS0 = items, NS 120 = properties
If my entity source for wikidata on my local install says that properties are
in NS 120, then links will be created locally for NS 120, which is infact where
items are locally.
This results in "Item:P31" instead of ":P31"

This fact is now also documented in the fed props docs.

This was noticed while investigating T258564 (the linked bug) but
does not solve that task, but was in the way of solving it.

Bug: T258564
Change-Id: Id5e6bd7761aeaa49fa4c749aa143ca7bd3e848ab