Fix CachingFallbackLabelDescriptionLookup failing in edge-cases
This patch is not just silencing the error, it's a proper fix. See,
this class does not know anything about the TermLanguageFallbackChain
it gets. There is no guarantee given about the number of elements in
the chain. It *should* contain at least the interface language, but
as said, no guarantee.

Returning null is fine. It's documented to be possible, and all
callers are fine with null.

Bug: T259744
Change-Id: I38acd68e8273db086ecf18263e8712f096cc0220