Fix SqlChangeStoreTest after code rearrangement
The Lib EntityChange class no longer has an setMetadataFromRC() method –
that is now in RepoEntityChange, to make the Changes library independent
from MediaWiki. This means that SqlChangeStoreTest can no longer use the
method to prepare a change; we can simply set the appropriate fields and
metadata directly, though. (setMetadataFromRC() is tested explicitly via
RepoEntityChangeTest, so we shouldn’t lose any test coverage with this.)

Bug: T259814
Change-Id: I294c1f6fc1617b7a51e03c2d350cf68e8267202f
Depends-On: I0df26d6af56ed18eac836929fd469b4a637d4902