Make qunit RL module of repo use package modules
Lots of repo qunit tests are written as package module [1] and load
production code using package modules making it impossible to migrate
them to QUnitTestModule without that being package module (you can't use
scripts and packageFiles in the same module).

Note that since package files only loads the first file and leaves the
responsibility of loading the rest to the first file, we need to make an
index.js to load all of tests (and files they need) which is not super
nice but I don't think there is any other way.

[1] https://www.mediawiki.org/wiki/ResourceLoader/Package_modules

Bug: T259697
Change-Id: I6e45767e045f9466210a13fa1eeb45da2e46a8b1