FP: Use correct interwiki for fed props entity source
In order for the formatting of summaries to correctly work the
interwiki used in the entitysource must be defined as it is checked.
wd was used but is not defined by default, wikidata however is.

This change also slightly alters the fed props docs to reflect the
fact that the interwiki does need to be correct, and also makes the
example look more realistic.

The LinkRenderedEndHook was once again touched as apparently the hook
can also parse links such as "wikidata:Special:EntityPage/Q123" as being
in the "main namespace" with ns id 0 for the local wiki, even though to
us humans if is obviously in the -1 special NS on the remote wiki.

As far as I can tell the exact case that I touching here is already tested.
So I am a bit confused currently as to why the tests did not fail.

Bug: T258564
Change-Id: I5d7199ceb26d2b449e0b9b70cc43219200442076