phpdoc: use callable instead of callback
callable is the real reserved keyword, and there are only
2 cases where it is wrong.
https://www.php.net/manual/en/language.types.callable.php

callback is a pseudo-type
https://www.php.net/manual/en/language.pseudo-types.php

This doesn't seem that important, as they are documented
as being "the same", however as we more toward using types
there are apparently some subtle differences.

For example using callback as a param type can result in:
"
PHP Fatal error:  Uncaught TypeError:
Argument 1 passed to method() must be an instance of callback, instance of Closure given
"

Using callable instead of callback here avoids the fatal.

Change-Id: I5b1ecf5e0c74db3ff75d3007429a4c873c47dcf5