composer: psr/log ^1.1.3
Both MediaWiki and lib/changes require 1.1.3 so update this here too.

Would probably be good to have some automation making package bumps like
this as we start to have more composer.json files.

The question also arises for some of these packaged that are already required by
mediawiki, is there any point in duplicating the requirement here?
Probably? But again all of this needs more automation..

Change-Id: I90996c9bb8cafa4b23f52b2b35a62f7fbc2ac177