FP: Use tmpfs for test system's WDQS /tmp folder
The WDQS docker container creates a new temporary jetty folder in /tmp
every time it is (re-)started. In order to prevent the file system from
filling up over time, the /tmp folder is now mounted as tmpfs via
docker-compose.yml. This way it is kept in memory and will be removed
every time the container stops.

Bug: T260348
Change-Id: I6a1d94eb1f4912332da9fc076d85cf21bab343c0