Fix Term Fallback chains being empty
This could happen in cases when the given language is not an valid
built-in language code and thus the core fallback method returns an
empty array in
https://github.com/wikimedia/mediawiki/blob/a1ca5339e368c2396747386c7bf2275756d8af36/includes/language/LanguageFallback.php#L99-L101
Examples could be an emoji or ⧼Lang⧽. Since these languages are also not
valid term languages they are subsequently filtered in this
TermLanguageFallbackChain constructor, leaving the chain empty.

However, these chains should never be empty. It can be argued that they
should actually end in the wiki's content language instead of a
hard-coded English. But, hard-coding English is defacto the current behavior
of the main core fallback method mentioned above in
https://github.com/wikimedia/mediawiki/blob/a1ca5339e368c2396747386c7bf2275756d8af36/includes/language/LanguageFallback.php#L104-L106

Ensuring that it falls back to the wiki's (entity source's?) content
language will be a separate task. Maybe T260431

Bug: T259783
Bug: T259779
Change-Id: Ib43866b21f5e06764927557c697fba695abf6893