More consistent handling of cached URLs
EntityDataUriManager::getCacheableUrls() is split into two methods (and
gains a $revision parameter): the first method, getCacheableUrls(),
returns all canonical URLs for which caching may be enabled; the second,
getPotentiallyCachedUrls(), returns the internal versions of those URLs,
to be sent to the cache for purging if necessary. (In production, there
is no difference between canonical and internal URLs.)

This pair of methods is now fully responsible for determining whether a
URL should be cached or not. The logic which previously resided within
EntityDataRequestHandler::outputData(), to only cache responses for
specific revisions (see change I38175f43e8, commit 3856c7c324), is moved
into those methods; EntityDataRequestHandler now only checks if the
request URL is one of the cacheable URLs, as determined by the
EntityDataUriManager.

This means that the requirements on cacheable URLs are much stricter
now. Previously, any URL was in practice cacheable as long as it
included a revision parameter, and it could include arbitrary garbage as
well; now, we have to be able to *construct* the URL, otherwise we could
not purge it when necessary. We could try to construct the URL based on
the Title methods; however, that would mean the URLs would have to look
like this on Wikidata:

https://www.wikidata.org/w/index.php?title=Special:EntityData/Q42.json&revision=123

There are virtually no requests for such URLs; instead, commonly
requested URLs look more like this:

https://www.wikidata.org/wiki/Special:EntityData/Q42.json?revision=123

There is no MediaWiki method to construct such URLs – as soon as there
are any query parameters, MediaWiki builds the URL using the script path
(/w/), not the article path (/wiki/$1); see also T14619. Therefore, a
new configuration setting is added which specifies URL path patterns
that can be cached. By default, it includes the two patterns

/wiki/Special:EntityData/Q42.json?revision=123
/wiki/Special:EntityData/Q42.ttl?flavor=dump&revision=123

(but using $wgArticlePath), so that requests for these URLs can also
still be cached (and purged). Additional patterns can be added as
necessary, either in Wikibase.git (if they’re useful for most installs)
or in production config (if the requests are specific to Wikidata).

Change-Id: I217d38e89f28a5b2e3c4828f637c440da94b6dc8
Depends-On: I85269e3c1bb0d211f8d923ee4682044728ce9542