SECURITY: Add EntityDataPurger
This subscribes to the ArticleRevisionVisibilitySet hook and purges
affected Special:EntityData URLs from the web cache. For simplicity, it
does this regardless of whether the visibility change affected the
cached data (currently, the cached data does not include the user name
or comment associated with the revision, but who knows, maybe that’ll
change in the future).

This is implemented as a second hook handler, independent of the already
existing ArticleRevisionVisibilitySetHookHandler, because the two do
very different things and would share almost no code even if they were
merged into one class.

Bug: T260349
Change-Id: Ief8aba38205187fe969e8d87471919b2b258c615
Co-Authored-By: Itamar Givon <itamar.givon@wikimedia.de>