Add EntityDataFormatProvider to WikibaseRepo
The call to setAllowedFormats() was already duplicated in two places,
and I almost forgot to copy it when creating an EntityDataFormatProvider
in a third place. Let’s make a method for it in WikibaseRepo.

Also, ViewEntityAction had some similar code that can also use the new
getter (instead of accessing the entityDataFormats directly, without
setting it on the EntityDataFormatProvider).

Change-Id: I45e682641951d8a05ff6e5403a3417b4e620dcaf