EntitySavingHelper: allow ArrayAccess object status values
As of I53e39be59228ac5a57f34d51d733d1647331889c the `value` field in
Status objects may be an `ArrayAccess` object instead of a primitive
Array. This broke the error handling code in `EntitySavingHelper` because
it checks the status `value` field with `is_array` which returns false
for such objects.

The `is_array` checks can simply be removed, as `isset` is safe to be
used with an offset on a variable even if it is not array-like.

Bug: T260869
Change-Id: I32943ef4d2f386a6f5afe64561cebadb444c0f2d