Fix watch-on-edit settings on grant-limited requests
If a user made an entity edit that should add the affected page to the
user’s watchlist, but the user did not have the 'editmywatchlist' right
in the current session (either because the wiki is configured to deny
some users that right, or because the session has limited grants, e. g.
through a bot password or OAuth), then previously the page would not
have been added to the watchlist. MediaWiki avoids this by passing the
IGNORE_USER_RIGHTS flag when necessary, allowing some watchlist updates
regardless of user rights. Do the same in Wikibase. (The flag is needed
not just when modifying the watchlist, to bypass the 'editmywatchlist'
right check, but also when testing whether the page is already being
watched, since 'viewmywatchlist' is also a right that may be missing.)

WikiPageEntityStore::isWatching is called in two places and both benefit
from adding the User::IGNORE_USER_RIGHTS flag: Both
\Wikibase\Repo\UpdateRepo\UpdateRepoJob::saveChanges and
\Wikibase\Repo\EditEntity\MediawikiEditEntity::getWatchDefault try to
get the watch status of a page in order to not change it. However, since
the 'editmywatchlist' right is still required to _unwatch_ a page,
having WikiPageEntityStore::isWatching default to false is not harmful
either. So adding the flag to ignore the right is done help with
understanding and following the code, it will not affect behavior.

Bug: T217144
Change-Id: Ia24715bd69f6fc7c917666448f8af80662d67f6a
(cherry picked from commit 0adc27077904c9c780649544f047c69f47b49407)