OutputPageBeforeHTMLHookHandler: run for Special:Undelete
This introduces a new OutputPageEntityViewChecker that checks whether an
OutputPage object is likely to contain an entity view. The two guards
residing in OutputPageBeforeHTMLHookHandler and OutputPageEntityIdReader
that now use OutputPageEntityViewChecker previously did not take cases
into account where the OutputPage was not an entity article page, but
still contained an entity view, as it is the case for Special:Undelete.

Bug: T259541
Bug: T228147
Bug: T261511
Change-Id: Ibbe60f5af2302afd3e393238eebb2ce011692e02