Parse string values via PHP, not JS
We parse all other data types by sending them to the PHP parser via the
API – only the string data type was an exception. (Note that even other
data types with the string data *value* type, like external ID and URL,
use the PHP parsers.) Presumably, this was done under the assumption
that string parsing is simple, to the point that the JS StringParser [1]
is basically a no-op. However, the PHP StringParser [2] is slightly more
complex: it uses a StringNormalizer interface to normalize the string,
and in Wikibase, the WikibaseStringValueNormalizer implementation [3]
uses a StringNormalizer class [4] (not the same as the StringNormalizer
interface) to trim whitespace and normalize the string to NFC. Instead
of trying to reimplement this trimming and normalization in JavaScript
(the PHP code looks reasonably nontrivial), let’s just stop pretending
that we have a JS implementation equivalent to the PHP implementation,
and parse all string values via the API, just like all other data types.

[1]: https://github.com/wmde/DataValuesJavaScript/blob/59086c58bdb1b95a1c653b9cc4b7c1ebf69f117c/src/valueParsers/parsers/StringParser.js
[2]: https://github.com/DataValues/Common/blob/f4acd2add0228174298d423473d57afb86add4b4/src/ValueParsers/StringParser.php
[3]: https://gerrit.wikimedia.org/g/mediawiki/extensions/Wikibase/+/b041eea8633ee566278fe7d881c6ec40d00c9f87/repo/includes/Parsers/WikibaseStringValueNormalizer.php
[4]: https://gerrit.wikimedia.org/g/mediawiki/extensions/Wikibase/+/b041eea8633ee566278fe7d881c6ec40d00c9f87/lib/includes/StringNormalizer.php

Bug: T261071
Change-Id: I6548c0149a12449ecd4c941e69bf9e0267b7ebf2