Stop sending JS parsers to clients
The StringParser is unused since the parent commit (change I6548c0149a);
the BoolParser, FloatParser and IntParser have never been used, as far
as I could tell. (We stopped registering them in PHP in 2014 – see
commit 91f9758dae, change I98a4dd29c9.)

The ValueParser base class and the NullParser implementation are still
used in our JavaScript code, so keep those.

Change-Id: I195f39d8049abaa87c685d8616abcca7380c0586