Restore MakeRowIDsBig.sql update for ips_row_id
This update was removed in change I6b7c1700cb (commit 310977066c), but
the part of it about the wb_items_per_site table should still be kept.

This will cause the update to run again even on wikis where the field is
already a BIGINT, but that’s okay.

Change-Id: Ic57641d2cbf1624e82afe5883f19860297c4bc4a