Remove multiple-term-store logic from WikibaseRepo
ItemHandler and PropertyHandler now get a single EntityTermStoreWriter
injected, instead of a list of them, and the previous logic in
WikibaseRepo to build those lists is boiled down to the parts that are
still relevant in a post-wb_terms world. This makes the classes
ByIdDispatchingItemTermStoreWriter and NullItemTermStoreWriter unused;
I’ll remove them separately.

Since the old tests for the affected methods in WikibaseRepoTest no
longer apply, we now instead test that, without local items/properties,
we don’t try to build term stores for those entity types.

Bug: T264602
Change-Id: I088e34cc6018320ef61f88ab0085e1ab5458b9b7