Use newer wikibase codesniffer that includes mediawiki cs 31
This patch currently includes the fixes done by phpcbf. Those fixes
mainly consist of changing `$this->assertEquals( 1, ` into
`$this->assertSame( 1, ` as a lot of data structures could be type
wrangled into the integer 1, e.g. true. While this problem is less
pronounced for integers other than 1 and 0, there is still rarely a good
reason to use the less type safe assertion method. Therefore I change
all instances of `$this->assertEquals(` to `$this->assertSame(` in files
that were touched for this reason anyway.

Bug: T264771
Change-Id: I73bb1b03c50e5e1419b25625ccfb49400822c307