Remove MatchingTermsLookup::getTopMatchingTerms()
In the only remaining real implementation, it’s the same as
getMatchingTerms(). This is actually a violation of the method’s
contract – it’s only supposed to return one result per entity ID – but
it turns out nobody relies on this behavior, so it’s safe to remove.

The MatchingTermsLookupSearchInteractorTest *does* rely on the filtering
of entity ID (which DatabaseMatchingTermsLookup did not implement, but
MockMatchingTermsLookup did), so temporarily add such a filter directly
in the test. I’ll update the test case to expect duplicate results in a
separate commit.

Bug: T265493
Change-Id: I37fe78968ded4d7d98294c0fdb38b6d8ccb65cd1