Make getTextForSummary less forgiving.
After discussing a way forward with refactoring this method with Lucas,
Michael and Pablo we have considered the following options.

- Making this method abstract and forcing its subclasses to implement it
to get rid of the dependency on LabelsProvider from EntityContent.

This was tried in 634006 but was deemed undesirable.

- Adding some kind of Summary formatter that would be injected into each
of the EntityContent implementations and move the responsibility of
creating the summary for the specific entity type there.

This has not yet been tried but reviewing the possibility of injecting
this into each EntityContent implementation this was also deemed
undesirable.

Therefore we made a compromise where we decided to make the original
method less forgiving and instead of letting an EntityContent with a
entity that is not a LabelsProvider use this method quietly return an
empty string this method should now throw an exception. LexemeContent is
the exception in this case and should implement its own method as
suggested in Ibe0b46ed9c81ff9be2be2e8c7201d76f5635c97e.

Change-Id: Idf6c42f4c25598f50f955d8c33b6873865fa0ff2
Depends-On: Ibe0b46ed9c81ff9be2be2e8c7201d76f5635c97e
Bug: T263435