FormatterCacheFactory: use a factory for all services
This moves (sub)service creation out of FormatterCacheFactory to make it
testable and allow for a potential reconfiguration at some point -
unfortunately the type hints are very class-based at the moment.
Ideally, all those hand-crafted, zero-logic factory methods would be
replaced by a configurable container builder [1] and/or autowiring [2].

SimpleCacheWithBagOStuff is losing its `final` again, as testability
without black magic is a "reason not to do so" and 6204535 states no
other argument why it could be important to have it.

[1]: https://symfony.com/doc/current/components/dependency_injection.html#setting-up-the-container-with-configuration-files
[2]: https://symfony.com/doc/current/service_container/autowiring.html#an-autowiring-example

Bug: T265983
Change-Id: I9ef6250902ea4752e7d58544d85e5cc6b2b3d337