Add implicit usage on description in content language
Introduce the ImplicitDescriptionUsageLookup class, which wraps an inner
UsageLookup to add an implicit usage on the description in the content
language of the linked item of a page. The class is used if the
enableImplicitDescriptionUsage client setting is enabled (disabled by
default). usagetracking.md is updated to document implicit usage too.

Bug: T265897
Change-Id: I66b1092bb67163a930c579547aca50837453ee58