GeoData: Explicitly set coordinates to ParserOutput
Transitioning ParserCache to JSON will require all extension
data to be serialized into JSON before adding to ParserOutput.
The old pattern of using CoordinatesOutput::getOrBuildFromParserOutput
was relying on setting the new coords to extensionData insternally,
and then modifying the mutable object. This will not be allowed
anymore, since the extension data should be serialized when setting.

Thus, the code needs to explicitly set the updated CoordinatesOutput
to ParserOutput extension data.

Bug: T266248
Depends-On: Iaab041b13e448c0a42735d5d028a625a016b44cc
Change-Id: Iba8618236c39c95d27fa018f097e623477039feb