Try to clarify title usage
To me, “the entity’s local page name” suggested that this usage tracked
the title of the entity on *its* local wiki, e. g. the item ID (which
would change e. g. when merging entities). But this is not the case: the
usage refers to the title of the local page linked to the entity. Try to
clarify this a bit.

Change-Id: I5d406fa630771e7d021104f0e56ca6dc9cfb7f58