Remove pointless LinkBatch in ChangeHandler
Running a LinkBatch is useful if we have a set of titles that were
created from text form, whose page ID and other data may not have been
loaded yet. However, in this case we actually have a set of titles that
were loaded by page ID, and in that case the data is loaded already, so
running the LinkBatch is, as far as I can tell, just a waste of time.
(The LinkBatch could in theory detect that the data is already loaded
and skip the database query, but currently that doesn’t happen, see
T266494.)

This reverts commit 33c97f9abb74ec0735c96c0852f0965de6e6c3db.

Bug: T266499
Change-Id: Ie9480757dcd6386d6f9e7eefcf40f8a6014d5c08