Don't write referenced-entities into ParserOutput.
ReferencedEntitiesDataUpdater was writing a list of EntityId objects
into the 'referenced-entities' extension data field in ParserOutput.
ParserOutput no longer supports arbitrary objects as extension data.
Since 'referenced-entities' appear to be unused, the easiest way to
resolve this is to simply no longer write the referenced entities to
ParserOutput. (Note: the method EntityParserOutputGenerator::getEntityInfo,
which used to need this information, no longer exists).

If the list of referenced entities is needed in ParserOutput in the
future, the EntityIds can simply be converted to their string
representation.

Bug: T266263
Change-Id: Ifaff8b3e0fbbc1b4e350ddc648fdcc1f0b1f4c54